Adds missing checks, and refactors verify_deployment to its own file
